+++ R/head.R
@@ -7,33 +7,42 @@
 #' @param x A `gtsummary` object.
 #' @param n Number of rows to return. Default is `6L` as default in [utils::head()] and
 #'   [utils::tail()].
-#' @param ... Additional arguments passed to [utils::head()] or [utils::tail()].
+#' @param ... Additional arguments passed to [utils::head()], [utils::tail()], or `[`.
 #'
 #' @return A `gtsummary` object with only the first or last `n` rows in `table_body`.
+#' @keywords internal
+#' @name head
 #'
 #' @examples
-#' # head() example
-#' trial |>
-#'   tbl_summary(by = trt) |>
-#'   head(n = 2L)
+#' tbl <- tbl_summary(trial, by = trt, include = grade)
 #'
-#' @keywords internal
+#' # Example 1 head() ---------------------------
+#' head(tbl, n = 2L)
+#'
+#' # Example 2 tail() ---------------------------
+#' tail(tbl, n = 2L)
+#'
+#' # Example 3 `[`() ---------------------------
+#' tbl[1:2,]
+NULL
+
+#' @rdname head
 #' @export
 head.gtsummary <- function(x, n = 6L, ...) {
   x$table_body <- head(x$table_body, n = n, ...) # data.frame head
   x
 }
 
-#' @examples
-#' # tail() example
-#' trial |>
-#'   tbl_summary(by = trt) |>
-#'   tail(n = 2L)
-#'
-#' @rdname head.gtsummary
-#' @keywords internal
+#' @rdname head
 #' @export
 tail.gtsummary <- function(x, n = 6L, ...) {
   x$table_body <- tail(x$table_body, n = n, ...) # data.frame tail
+  x
+}
+
+#' @rdname head
+#' @export
+`[.gtsummary` <- function(x, ...) {
+  x$table_body <- x$table_body[...]
   x
 }