+++ NEWS.md
@@ -1,10 +1,16 @@
 # gtsummary (development version)
 
+* Added AI chat bot to homepage, powered by `kapa.ai`.
+
+* Added `head()` and `tail()` S3 methods for `gtsummary` objects. (#2335) 
+
 * Updated `tbl_ard_hierarchical()` to pre-process non-standard hierarchical ARDs so sorting can be applied. (#2308)
 
 * Added theme element `"assign_summary_type-arg:cat_threshold"` for greater control over default summary types.
 
 * When a data frame is passed in the `tbl_summary(percent)` argument, the headers are now tabulated with this data frame. (#2322)
+
+* Corrected handling in `tbl_summary()` of character vectors that contains `"NULL"` values. (#2345)
 
 # gtsummary 2.4.0