+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added AI chat bot to homepage, powered by `kapa.ai`.
 
+* Adding functions `tlb_ard_strata()` and `tlb_ard_strata2()`. (#1852)
+
 * Added `head()` and `tail()` S3 methods for `gtsummary` objects. (#2335) 
 
 * Updated `tbl_ard_hierarchical()` to pre-process non-standard hierarchical ARDs so sorting can be applied. (#2308)