+++ NEWS.md
@@ -18,6 +18,8 @@
 
 * Fixed bug in `sort_hierarchical()` causing variable labels to disappear from table when not all variables are in `include`. (#2332)
 
+* Added the `"add_overall.tbl_svysummary-arg:col_label"` theme element to control the default value for `add_overall.tbl_svysummary(col_label)` argument. (#2354)
+
 # gtsummary 2.4.0
 
 ### New Features and Functions