+++ NEWS.md
@@ -1,4 +1,4 @@
-# gtsummary (development version)
+# gtsummary 2.5.0
 
 ### New Features and Functions